Move check_timelocks to a separate module for transaction verifier
